feat: add serialized test cases
Adds support for serialized test cases,
which take a `SerializedTestContext` as parameter,
a type alias for `TestContext<true>`.
The context is now composed of a `bool` const generic
(defaults to `false` to not change the syntax) which
specify if the context should run in a serialized context.
As a result,
`SerializedTestContext` provides the same functions as `TestContext`,
plus functions which should be run in a serialized context (e.g `as_user`).