Converting NFSv4 granular permissions tests
But name it "nfsv4acl" instead of "granular".  In the future, there may
be a "posix1eacl" module too.

Don't run the NFSv4 ACL tests on UFS, for now, due to a FreeBSD bug.

I did not convert any tests that involve combinations of
ACL_DELETE_CHILD on a directory and ACL_DELETE on files within that
directory.  The NFSv4 RFC does not specify what should happen in that
case.  The existing tests seem to have simply asserted ZFS's behavior at
the time the tests were written, which may not have been well
thought-out.  For example, ZFS requires ACL_WRITE_DATA to delete
subdirectories, but ACL_APPEND_DATA to create them.  Finally, the only
documentation I could find that _does_ deal with combinations of
ACL_DELETE_CHILD and ACL_DELETE is the Linux nfs4_acl(5) man page, which
says that the behavior is file-system-dependent.

Issue #14